[fit] 폴더명 수정, readMe 수정
