+++ src/pages/Layout/Layout.tsx
@@ -1,5 +1,5 @@
-import Footer from '@/components/footer/Footer';
-import Header from '@/components/header/Header';
+import Footer from '@/components/Footer/Footer';
+import Header from '@/components/Header/Header';
 import { Outlet } from 'react-router-dom';
 import styled from 'styled-components';
